refactor: move some Settings components to @/settings
